# docs/openapi.yaml
openapi: 3.1.0
info:
  title: discord-support-hub API
  version: 0.1.0
  summary: API-driven provisioning and governance of isolated private Discord support spaces.
  description: |
    The control surface (FR-11) for discord-support-hub, consumed by the backoffice
    (no human UI in v1). Postgres is the authorization source of truth; Discord is a
    projection the bot reconciles (three-layer truth model).

    Conventions:
      * **Mutating endpoints are asynchronous.** They validate, persist the *desired*
        state, enqueue a job, and return `202 Accepted` with a `job` handle. Poll
        `GET /jobs/{jobId}` for completion. (NFR-2: latency decoupled from Discord rate limits.)
      * **Read endpoints** are served from cache (TTL + write-invalidation) and return `200`.
      * **Authorization always resolves against Postgres**, never the Discord role (NFR-13).
      * **Idempotency:** mutating requests accept an `Idempotency-Key` header; a repeated key
        replays the original response instead of re-enqueueing (NFR-3).
      * **Role-per-merchant access.** Provisioning a space auto-creates a Discord role for the
        merchant and grants that role VIEW+SEND on the channel. A collaborator acquires access by
        joining through the merchant's native invite-with-role link (created once by the operator
        in the Discord client and stored via `PUT /merchants/{merchantId}/invite`), which the hub
...
  license:
    name: Apache-2.0
    identifier: Apache-2.0
...
servers:
...
security:
  - ServiceApiKey: []
...
paths:
...
  /merchants:
    post:
      tags: [Merchants]
      operationId: registerMerchant
      summary: Register a merchant (synchronous).
      description: |
        Records a new merchant in Postgres - the authorization source of truth - and returns
        `201` with the created `Merchant`. This is a synchronous Postgres write with **no Discord
        side-effect** (a merchant owns a space, but the space is provisioned separately via
        `POST /merchants/{merchantId}/channels`); there is no job to poll.

        The `id` returned here is the merchant UUID used as `{merchantId}` everywhere else in this
        API (provisioning, filters, audit). The backoffice should persist it.

        `external_ref` is the backoffice's own stable identifier and is **unique**: a second
        registration with the same `external_ref` returns `409` (it does not create a duplicate).
        This makes registration safe to retry on the `external_ref` natural key.

        Control-plane authority required (backoffice-scoped service key, NFR-13).
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterMerchantRequest'
      responses:
        '201':
          description: Merchant registered. The `id` is the merchant UUID used as `{merchantId}`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Merchant' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: |
            A merchant with this `external_ref` already exists (UNIQUE constraint), or an
            `Idempotency-Key` was reused with a different body.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
    get:
      tags: [Merchants]
      operationId: listMerchants
      summary: List merchants (cursor pagination).
      description: |
        Returns a page of merchants ordered by `created_at` ascending, consistent with the other
        list endpoints (`GET /channels`, `GET /directory`, `GET /audit`). Cursor pagination:
        pass the returned `next_cursor` back as `cursor` to fetch the next page; a `null`
        `next_cursor` means the last page. Control-plane authority required.
      parameters:
        - name: is_active
          in: query
          required: false
          description: When set, filters to active (`true`) or inactive (`false`) merchants.
          schema: { type: boolean }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Page of merchants.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Merchant' }
                  next_cursor: { type: [string, 'null'] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /merchants/{merchantId}:
    get:
      tags: [Merchants]
      operationId: getMerchant
      summary: Merchant detail by UUID id.
      description: |
        Returns the merchant identified by the UUID `{merchantId}` (the `id` returned by
        `POST /merchants`). Returns `404` with a `{code, message}` body when no merchant has that
        id (including when `{merchantId}` is not a well-formed UUID). Control-plane authority required.
      parameters:
        - $ref: '#/components/parameters/MerchantId'
      responses:
        '200':
          description: Merchant detail.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Merchant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    ServiceApiKey:
      type: http
      scheme: bearer
      description: |
        Opaque service API key issued to the backoffice (AuthZ Layer A). Sent as
        `Authorization: Bearer <key>`. Stored hashed at rest; instantly revocable. The future POC
        frontend uses a separate session principal (not this long-lived key) behind the same
        authorization layer.
  parameters:
    MerchantId:
      name: merchantId
      in: path
      required: true
      description: |
        The merchant **UUID `id`** (the `id` returned by `POST /merchants`), NOT the
        backoffice `external_ref`. A value that is not a well-formed UUID, or one that matches
        no merchant, resolves to `404`.
      schema: { type: string, format: uuid }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-supplied key making a mutating request idempotent (NFR-3). A repeated key with the
        same body replays the original response; a repeated key with a different body returns 409.
      schema: { type: string, maxLength: 255 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid service API key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized (resolved against Postgres, NFR-13).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    RegisterMerchantRequest:
      type: object
      description: Body for POST /merchants. `external_ref` is the backoffice's stable id and is unique.
      required: [external_ref, name]
      properties:
        external_ref:
          type: string
          minLength: 1
          description: |
            Stable identifier from the backoffice. Unique across merchants; a duplicate value
            returns `409`. Lets the backoffice address a merchant idempotently by its own id.
        name:
          type: string
          minLength: 1
          description: Human-readable merchant name.
        help_desk_url:
          type: [string, 'null']
          format: uri
          description: Optional per-merchant help-desk link (FR-15 static parameterization).
    Merchant:
      type: object
      description: |
        An external customer. Owns exactly one space (1:1) and one Discord merchant role (1:1).
        Stores the per-merchant invite-with-role link used to onboard its collaborators.
      required: [id, external_ref, name, is_active, created_at]
      properties:
        id:
          type: string
          format: uuid
          description: Merchant UUID. This is the `{merchantId}` used elsewhere in the API.
        external_ref:
          type: string
          description: The backoffice's stable id (unique).
        name: { type: string }
        help_desk_url:
          type: [string, 'null']
          format: uri
        invite_link:
          type: [string, 'null']
          format: uri
          description: |
            The merchant's stored native invite-with-role link (set via PUT /merchants/{merchantId}/invite).
            Null until the operator stores it; :send-invite is rejected while null.
        invite_link_set_at:
          type: [string, 'null']
          format: date-time
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code, e.g. validation_error, forbidden, not_found, idempotency_conflict.
        message: { type: string }
        details:
          type: [object, 'null']
          additionalProperties: true